Initial setup character-list
